[CREATE] TimeAddModalViewController created
